29/01 v1

+++ 1_Data_Access_2/7_Hibernate/Parte3_ConSpring/hibernate/src/main/java/org/example/nachoHibernateConSpring/service/PatientService.java
@@ -1,12 +1,15 @@
 package org.example.nachoHibernateConSpring.service;
 
+import org.example.nachoHibernateConSpring.dao.model.Credential;
+import org.example.nachoHibernateConSpring.dao.model.MedRecord;
 import org.example.nachoHibernateConSpring.dao.model.Patient;
-import org.example.nachoHibernateConSpring.dao.repository.CredentialDAO;
-import org.example.nachoHibernateConSpring.dao.repository.PatientDAO;
+import org.example.nachoHibernateConSpring.dao.repository.*;
 import org.example.nachoHibernateConSpring.domain.error.MedicalRecordException;
 import org.example.nachoHibernateConSpring.domain.error.UsernameDuplicatedException;
 import org.example.nachoHibernateConSpring.domain.model.PatientUI;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -14,10 +17,18 @@ import java.util.List;
 public class PatientService {
     private final PatientDAO dao;
     private final CredentialDAO credentialDAO;
+    private final MedRecordDAO medRecordDAO;
+    private final MedicationDAO medicationDAO;
+    private final PaymentDAO paymentDAO;
+    private final AppointmentDao appointmentDao;
 
-    public PatientService(PatientDAO dao, CredentialDAO credentialDAO) {
+    public PatientService(PatientDAO dao, CredentialDAO credentialDAO, MedRecordDAO medRecordDAO, MedicationDAO medicationDAO, PaymentDAO paymentDAO, AppointmentDao appointmentDao) {
         this.dao = dao;
         this.credentialDAO = credentialDAO;
+        this.medRecordDAO = medRecordDAO;
+        this.medicationDAO = medicationDAO;
+        this.paymentDAO = paymentDAO;
+        this.appointmentDao = appointmentDao;
     }
 
     public List<PatientUI> getPatients() {
@@ -26,30 +37,74 @@ public class PatientService {
                 .map(Patient::toPatientUI).toList();
     }
 
+    @Transactional
     public int addPatient(PatientUI patientUI) {
-        // Check for duplicated username
-        if (credentialDAO.findByUsername(patientUI.getUserName()).size() > 0) {
-            throw new UsernameDuplicatedException("Username duplicated, it already exists");
-        }
+        try {
+            if (credentialDAO.findByUsername(patientUI.getUserName()).size() > 0) {
+                throw new UsernameDuplicatedException("Username duplicated, it already exists");
+            }
 
-        Patient savedPatient = dao.save(patientUI.toPatient());
+            Credential credential = new Credential();
+            credential.setUsername(patientUI.getUserName());
+            credential.setPassword(patientUI.getPassword());
 
-        if (savedPatient.getId() == 0) {
-            throw new MedicalRecordException("Unexpected error saving patient, no patient was saved, rolling back...");
+            Patient patient = patientUI.toPatient();
+            patient.setCredential(credential);
+            credential.setPatient(patient);
+
+            Patient savedPatient = dao.save(patient);
+
+            if (savedPatient.getId() == 0) {
+                throw new MedicalRecordException("Unexpected error saving patient, no patient was saved, rolling back...");
+            }
+            return savedPatient.getId();
+        } catch (DataIntegrityViolationException e) {
+            throw new MedicalRecordException("Data integrity violation while saving patient "+ e);
+        } catch (Exception e) {
+            throw new MedicalRecordException("Unexpected error while saving patient "+ e);
         }
-        return savedPatient.getId();
     }
 
     public void updatePatient(PatientUI patientUI) {
-        dao.save(patientUI.toPatient());
+        Patient existingPatient = dao.findById(patientUI.getId())
+                .orElseThrow(() -> new IllegalArgumentException("Patient not found"));
+
+        existingPatient.setName(patientUI.getName());
+        existingPatient.setBirthDate(patientUI.getBirthDate());
+        existingPatient.setPhone(patientUI.getPhone());
+
+        dao.save(existingPatient);
     }
 
+    @Transactional
     public boolean delete(int patientId, boolean confirm) {
-        // This way of making it makes it so that it will always call all those deletes even after checking if there are records, but it is ok, it is a small amount of data
-//        if (!confirm && !medRecordService.checkPatientMedRecords(patientId)) {
-//                throw new MedicalRecordException("Patient has medical records, cannot delete.");
-//        }
-        dao.deleteById(patientId);
+        Patient patient = dao.getById(patientId);
+        List<MedRecord> medRecords = medRecordDAO.findByPatientId(patientId);
+        if (!confirm && !medRecords.isEmpty()) {
+            throw new MedicalRecordException("Patient has medical records, cannot delete.");
+        } else {
+            try {
+                // Delete medications and medical records
+                medRecords.forEach(medRecord -> {
+                    medicationDAO.deleteByMedRecord(medRecord);
+                    medRecordDAO.deleteById(medRecord.getId());
+                });
+
+                // Delete payments
+                paymentDAO.deleteByPatientId(patientId);
+
+                // Delete appointments
+                appointmentDao.deleteByPatientId(patientId);
+
+                // Delete credentials
+                credentialDAO.deleteByPatient(patient);
+
+                // Delete patient
+                dao.deleteById(patientId);
+            } catch (DataIntegrityViolationException e) {
+                throw new MedicalRecordException("Error deleting patient data, rolling back...");
+            }
+        }
         return true;
     }
